+++ src/index.ts
@@ -1,3 +1,4 @@
 import VernaForm from './VernaForm';
-
+import { VernaContextProvider, useVerna } from './VernaContextProvider';
 export default VernaForm;
+export { VernaContextProvider, useVerna };