+++ src/types/JSONSchemaOrdered.ts
@@ -1,0 +1,7 @@
+import { JSONSchema7 } from 'json-schema';
+import type { UiSchema } from '@rjsf/core';
+
+export default interface JSONGlobalSchema {
+  schema: JSONSchema7;
+  uiSchema: UiSchema;
+}